+++ collections/src/main/kotlin/ktx/collections/maps.kt
@@ -1,4 +1,4 @@
-@file:Suppress("NOTHING_TO_INLINE")
+@file:Suppress("NOTHING_TO_INLINE", "LoopToCallChain")
 
 package ktx.collections
 
@@ -355,3 +355,47 @@ inline operator fun <Value> ObjectIntMap.Entry<Value>.component1() = key!!
  * @return [ObjectIntMap.Entry.value]
  */
 inline operator fun <Value> ObjectIntMap.Entry<Value>.component2() = value
+
+/**
+ * Returns a [GdxMap] containing the results of applying the given [transform] function
+ * to each entry in the original [GdxMap].
+ */
+inline fun <Key, Value, R> GdxMap<Key, Value>.map(transform: (Entry<Key, Value>) -> R): GdxMap<Key, R> {
+  val destination = GdxMap<Key, R>(this.size)
+  for (item in this) {
+    destination[item.key] = transform(item)
+  }
+  return destination
+}
+
+/**
+ * Returns a [GdxMap] containing only entries matching the given [predicate].
+ */
+inline fun <Key, Value> GdxMap<Key, Value>.filter(predicate: (Entry<Key, Value>) -> Boolean): GdxMap<Key, Value> {
+  val destination = GdxMap<Key, Value>()
+  for (item in this) {
+    if (predicate(item)) {
+      destination[item.key] = item.value
+    }
+  }
+  return destination
+}
+
+/**
+ * Returns a single [GdxArray] of all elements from all collections in the given [GdxMap].
+ */
+inline fun <Key, Type, Value : Iterable<Type>> GdxMap<Key, out Value>.flatten(): GdxArray<Type> {
+  val destination = GdxArray<Type>()
+  for (item in this) {
+    destination.addAll(item.value)
+  }
+  return destination
+}
+
+/**
+ * Returns a single [GdxArray] of all elements yielded from results of transform function being invoked
+ * on each entry of original [GdxMap].
+ */
+inline fun <Key, Value, R> GdxMap<Key, Value>.flatMap(transform: (Entry<Key, Value>) -> Iterable<R>): GdxArray<R> {
+  return this.map(transform).flatten()
+}