+++ collections/src/test/kotlin/ktx/collections/listsTest.kt
@@ -2,11 +2,11 @@ package ktx.collections
 
 import org.junit.Assert.*
 import org.junit.Test
+import java.util.LinkedList
 import java.util.NoSuchElementException
 
 /**
  * Tests general [PooledList] utilities.
- * @author MJ
  */
 class ListsTest {
   @Test
@@ -49,13 +49,12 @@ class ListsTest {
 
   @Test
   fun `should provide alias for compatibility with other LibGDX collections`() {
-    assertTrue(GdxList<Any>(NodePool) is PooledList<Any>)
+    assertTrue(GdxList(NodePool) is PooledList<Any>)
   }
 }
 
 /**
  * Tests [PooledList] implementation - a KTX LinkedList equivalent.
- * @author MJ
  */
 class PooledListTest {
   @Test
@@ -224,6 +223,49 @@ class PooledListTest {
     assertEquals(0, NodePool.free) // Nodes should not be returned to the pool.
   }
 
+  @Test
+  fun `should map elements into a new GdxList`() {
+    val list = gdxListOf(1, 2, 3)
+    val result = list.map { it * 2 }
+
+    assertTrue(result is GdxList)
+    assertEquals(3, result.size)
+    assertEquals(2, result.first)
+    assertEquals(6, result.last)
+  }
+
+  @Test
+  fun `should filter elements into a new GdxList`() {
+    val list = gdxListOf(1, 2, 3, 4, 5)
+    val result = list.filter { it % 2 == 1 }
+
+    assertTrue(result is GdxList)
+    assertEquals(3, result.size)
+    assertEquals(1, result.first)
+    assertEquals(5, result.last)
+  }
+
+  @Test
+  fun `should flatten elements into a new GdxList`() {
+    val list = gdxListOf(GdxArray.with(1), listOf<Int>(), LinkedList(arrayListOf(2, 3)))
+    val result = list.flatten()
+
+    assertTrue(result is GdxList)
+    assertEquals(3, result.size)
+    assertEquals(3, result.size)
+    assertEquals(1, result.first)
+    assertEquals(3, result.last)
+  }
+
+  @Test
+  fun `should map elements to lists and flatten them into a new GdxList`() {
+    val list = gdxListOf(1, 2, 3)
+    val result = list.flatMap { List(it) { "" } }
+
+    assertTrue(result is GdxList)
+    assertEquals(6, result.size)
+  }
+
   @Test(expected = NoSuchElementException::class)
   fun shouldThrowExceptionIfFirstElementIsRequestedButListIsEmpty() {
     PooledList(NodePool).first
@@ -387,5 +429,28 @@ class PooledListTest {
     assertEquals("[]", gdxListOf<String>().toString())
     assertEquals("[single]", gdxListOf("single").toString())
     assertEquals("[one, two, three]", gdxListOf("one", "two", "three").toString())
+  }
+
+  @Test
+  fun `should calculate distinct hash code for lists with same elements`() {
+    val list = gdxListOf("a", "b", "c")
+    val same = gdxListOf("a", "b", "c")
+    val different = gdxListOf("b", "c", "a")
+
+    assertEquals(same.hashCode(), list.hashCode())
+    assertNotEquals(different.hashCode(), list.hashCode())
+  }
+
+  @Test
+  fun `should properly implement equals`() {
+    val list = gdxListOf("a", "b", "c")
+
+    assertNotEquals(list, null)
+    assertNotEquals(list, "[a, b, c]")
+    assertNotEquals(list, gdxListOf<String>())
+    assertNotEquals(list, GdxArray.with("a", "b", "c")) // No common interface.
+    assertNotEquals(list, gdxListOf("a", "b"))
+    assertNotEquals(list, gdxListOf("a", "b", "c", "d"))
+    assertEquals(list, gdxListOf("a", "b", "c"))
   }
 }