+++ vis/src/main/kotlin/ktx/vis/widgets.kt
@@ -102,8 +102,7 @@ class KVisTree(styleName: String) : VisTree(styleName), TreeWidgetFactory {
   }
 }
 
-/** Extends [Tree] [Node] API with type-safe widget builders.
- * @author MJ */
+/** Extends [Tree] [Node] API with type-safe widget builders. */
 @VisDsl
 class KNode(actor: Actor) : Node(actor), TreeWidgetFactory {
   override fun <T : Actor> addActorToTree(actor: T): KNode {