+++ CHANGELOG.md
@@ -1,6 +1,17 @@
-#### 1.9.6-SNAPSHOT
+#### 1.9.6-b4
 
-- **[UPDATE]** Updated to Kotlin 1.1.2.
+- **[FEATURE]** (`ktx-collections`) Added `map`, `filter` and `flatten` extension methods that return LibGDX collections.
+- **[FEATURE]** (`ktx-collections`) `PooledList` now properly implements `hashCode` and `equals`.
+- **[FEATURE]** (`ktx-app`) Added `KtxGame`: **KTX** equivalent of LibGDX `Game`.
+- **[FEATURE]** (`ktx-app`) Added `KtxScreen`: adapter of the LibGDX `Screen` interface making all methods optional to implement.
+- **[FEATURE]** (`ktx-app`) Added `emptyScreen` utility method returning a no-op implementation of `Screen`.
+- **[FEATURE]** (`ktx-inject`) `Context` now implements `Disposable` and allows to dispose of all registered singletons and providers.
+- **[FEATURE]** (`ktx-inject`) Added `Context.remove` and `removeProvider` methods. Now providers for particular types can be removed without clearing the whole context.
+- **[FEATURE]** (`ktx-inject`) `getProvider`, `setProvider` and `clear` methods of `Context` are now open and can be overridden.
+
+#### 1.9.6-b3
+
+- **[UPDATE]** Updated to Kotlin 1.1.2-3.
 - **[UPDATE]** Updated to Kotlin Coroutines to 0.15.
 - **[CHANGE]** (`ktx-assets`) Static `AssetManager` instance container - `Assets` - was removed. All top level functions
 depending on the global `AssetManager` were removed.